Update
New mechanic: Auto-Drop. When doing aerials, you can hold Down to pass through soft platforms while attacking. For some aerials this is difficult, so for Gura's a.L she gets the HOLD_TO_DROP attack attribute which allows it to ignore platform if Light/Fierce are held during the attack.