Update
Changed the knockback on hammerhead DP to hit straight up.
Revised a bit on the priority system. The first 2 frames of Jabs (6 frame startup), will not hit opponents in attack startup. This means that Jabs will lose against any attack with 8 frames or lower startup. 8-frame startup is an important benchmark since it marks the beginning of the Held variation of certain moves. Gura's held g.uF, for instance, gain Superarmor on the 8th frame, and this change prevents it from being stuffed by Jabs if the two attacks start on the same frame.