Update
Removed "super jumps". Made Up + Jump an alternate input for shorthops, allowing shorthops to be done even on soft platforms. You can still shorthop with Down + Jump on solid platforms.
Set up foundations for "screenstops", which are very similar to hitstop. It will be used for certain Supers in the future.