Update
Made it easier to make conditional movement and attack data.
Gura can no longer do EX Sphyra (hammerhead DP) in the air. All frame 1 reversals will be limited to ground only (unless in special cases) in the future. Frame 1 reversals that can be done in the air are too absurd for 1 bar of EX.