Update
Added a Held Version to Gura's g.dF. It is a shorter stab, but if you sweetspot with it you will pull the opponent towards you instead of knocking them away. It can only be done once per combo though, and if you fail to sweetspot it is negative on hit.
Characters in Flinch Hitstun will now collide with other characters, it is only during Launch Hitstun that they will pass through other players.