Update
Gura's g.L can now be held for a 3-stab series instead of the normal 2. However, the 3rd hit cannot be chained out of on block, so don't use it against blocking opponents.
Changed input of Gura's command grab to Release Special + Down + Fierce. The reason I used Aux before is to prevent people from impulse-kara-cancelling into it, but now I have made it so that if you are holding down left/right you cannot use the command grab. In addition, if you had released left/right within 4 frames before using the command grab, it will not come out. This should prevent impulse-kara-cancelling into the command grab. To make up for being harder to do walk up throw, the hitbox for the command grab is improved a little.