Update
Changed how held input for held version of attacks works. In order for the held version of a move to be used, at the 8th frame after startup the game checks if the attack button is held down and is not tapped within the last 7 frames. This makes it harder for the held version of attacks to come out accidentally.
Readjusted the water effects for Harma. I think I finally understand how to sprite water effects.