Update
Converted dictionary keys used for storing attack data and doing hit detections into enumerations, after some testing showing that the string finding in the old version seem to be slowing down the game.
Modified the physics system, if the "target" position of a moving object is clear of obstacles, it will change it's position straight there instead of doing collision checking for each moved pixel. It can cause issues for very fast objects and thin walls, but the code runs much faster now when there are many objects on screen.