Update
Remade the physics engine in an attempt to reduce strain when many objects are on screen, using a "linear movement" method instead of the "pixel-by-pixel" method of the old engine.
...it didn't work. Performance of game is still terrible when many objects are on screen, with an upper cap of around 40 projectiles. Oh well.
I was hoping that I will be able to use holoVersus engine for an original game I have in mind. But I guess I will have to stick with Godot's built-in engine that is capable of supporting hundreds of projectiles. At least I should be able to reuse some pixel SFX from holoVersus, I guess.