Update
Before this update, multi-hit moves with different knockback on non-last hits vs last hits tend to behave oddly if they fail to score all hits. I finally figured out a solution: adding a "last-hit time marker" to the moves. If the multi-hit move landed with the animation time equal or later than the last-hit time marker, the hit is registered as a "last hit" and the ignore time is set to 999.
This also adds support for future multi-hit moves that only have hitboxes on certain intervals and can only land 1 hit per interval.

The character idle sprites in the character select screens now have the same ground level.