Update
All game music are now fully external, located in a GameMusic folder outside the .exe file. This means adding and removing music is as simple as drag and drop.
If there are multiple OGG files in a single folder, a random one will be picked.

You can now cancel jumpsquat into any attack if Up is held and Jump is not held, no longer requiring a UP_TILT indicator.
You can now cancel the 1st frame of Blocking into any attack if Block is held, for future Special Moves using the Special + Block command. You can actually parry for 1 frame before doing an attack now. I can remove this, but since it is pretty hard to pull off (require frame perfect input and timing) I will leave it in as a forbidden tech.