Update
Implemented Items.
Items are pretty much just consumable Assists. You can collect up to 3 of them (if you already have 3 the latest replace the oldest collected).
However, they are much easier to use than Selectable Assists. By pressing the Aux Button, the game automatically pick an Assist and Assist Attack based on the position of the opponent.
You can also use Items during hitstun, because Item Mode is not meant to be taken too seriously.
I will eventually implement Assists in Survival Mode as well, but there are too little Assists currently and Survival Mode will be redesigned in the future, so that wouldn't be done for a very long time.